move BaseDataModel to it's own file
